Reorganization, documentation, logging and messaging
Added a bit of logging in appointment and student controller, service and data model.

Added RestService interface for a couple of methods which are repeatedly used in all services.
Added Model interface - used e.g. in RestService.

Added DELETE, a couple of POST and PUT routes for appointments and students. Will also add for the other controllers and models.